requirements.txt
streamlit
openai-whisper
google-generativeai